Update index.jsx
Thank you for taking interest in writing a letter of recommendation!

The letter must be written, signed, and sent by the student's teacher, coach, or such. Maximum 1 page per letter.

The student will not be able to see the letter but will receive a notification once it is uploaded. The file must be a PDF and the size limit is 5 MB.